Support passing specific fields to `partial`
Support passing specific fields to `partial`. This restricts the fields
allowed to be missing during partial deserialization to those specified
in the value of `partial`.
This is particularly useful in api's when dealing with non client-generated
primary key fields (at creation), where only a specific field is allowed
to be missing, and you'd still like to use the same schema.