FIX GH#636: only is not bound by declared and additional fields on serializaton
Only field was not consistent with its treat of fields and declared
fields. To make it consistent fieldnames passed in as only not present
in Schema will be ignored. Two previous tests which were checking for
errors with invalid only fields were dropped as a result. Documentation
has also been updated to explicitly state that invalid fieldnames will
be ignored Documentation has also been updated to explicitly state that
invalid fieldnames will be ignored.