Fix type information for `__version_info__` to handle prereleases (#1925)
* Annotate the type of `__version_info__`

To avoid issues with `tuple[int, ...] + tuple[str, int]` on
pre-releases, which mypy flags when packaging data is available.

* Add `packaging` to mypy requirements

* Fix version_info type annotation with ignore

Although the cast to `tuple[int, int, int]` is acceptable to `mypy`
and will work on newer pythons, on older pythons it raises an error at
runtime. The reason is that `from __future__ import annotations` does
not "enable" newer typing features in the sense that other future
imports change language behaviors. Rather, it enables one feature:
deferred evaluation of annotations.

In the cast, the type is evaluated at runtime (even though it is not
used). As a result, the deferred evaluation no longer applies, and we
must either use `typing.Tuple` or remove the cast.

As it is simpler to read, this changes the assignments in `__init__`
to simply `type: ignore[assignment]`.